requirements.txt
apache-airflow
google-api-python-client
google-auth
pandas
numpy
matplotlib
seaborn
jupyter
python-dotenv
requests